backend/src/api/main.py
import os
import json
import tempfile
import traceback
from pathlib import Path
from typing import List, Dict, Any, Optional
import numpy as np
import pandas as pd
import pickle
import subprocess
import zipfile
from io import StringIO, BytesIO

from fastapi import FastAPI, File, UploadFile, HTTPException, Form, Response
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import FileResponse, StreamingResponse
from fastapi.staticfiles import StaticFiles
from pydantic import BaseModel
import uvicorn
from dotenv import load_dotenv

# Simple implementations to get the app running
from .utils.data_processing import DataProcessor

load_dotenv()

app = FastAPI(
    title="MBL IMUViz API",
    description="IMU Visualization and Inverse Kinematics Processing API",
    version="1.0.0"
)

# CORS middleware
origins = os.getenv("CORS_ORIGINS", "http://localhost:3000").split(",")
app.add_middleware(
    CORSMiddleware,
    allow_origins=origins,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Create directories
UPLOAD_DIR = Path(os.getenv("UPLOAD_DIR", "uploads"))
STATIC_DIR = Path(os.getenv("STATIC_DIR", "static"))
UPLOAD_DIR.mkdir(exist_ok=True)
STATIC_DIR.mkdir(exist_ok=True)
(STATIC_DIR / "imu").mkdir(exist_ok=True)

# Mount static files
app.mount("/static", StaticFiles(directory=str(STATIC_DIR)), name="static")

# Initialize services
data_processor = DataProcessor()

@app.get("/")
async def root():
    return {"message": "MBL IMUViz API is running"}

@app.post("/api/upload_sensor_mapping")
async def upload_sensor_mapping(file: UploadFile = File(...)):
    """Upload and parse sensor mapping file"""
    try:
        content = await file.read()
        content_str = content.decode("utf-8")
        
        mapping = data_processor.parse_sensor_mapping(content_str)
        return {"mapping": mapping}
    except Exception as e:
        raise HTTPException(status_code=400, detail=f"Error parsing sensor mapping: {str(e)}")

# Placeholder endpoints - will implement full functionality once basic structure works
@app.post("/api/upload_calibration")
async def upload_calibration(files: List[UploadFile] = File(...), task_id: str = Form("0")):
    return {"task_id": task_id, "offset": [0.0, 0.0, 0.0]}

@app.post("/api/upload_main_task") 
async def upload_main_task(files: List[UploadFile] = File(...)):
    return {"status": "received", "main_task_data": []}

@app.post("/api/run_ik")
async def run_ik(request: dict):
    return {"status": "success", "ik_results": {"joint_angles": {}}}

@app.get("/api/download_csv")
async def download_csv():
    return Response(content="No data", media_type="text/csv")

@app.get("/api/download_graphs_zip")
async def download_graphs_zip():
    return Response(content="No data", media_type="application/zip")

@app.get("/api/generate_video_stream")
async def generate_video_stream():
    return {"message": "Video generation not implemented yet"}

if __name__ == "__main__":
    uvicorn.run(
        "main:app",
        host=os.getenv("API_HOST", "0.0.0.0"),
        port=int(os.getenv("API_PORT", 8000)),
        reload=os.getenv("DEBUG", "false").lower() == "true"
    ) 