backend/pyproject.toml
[project]
name = "mbl-imuviz-api"
version = "0.1.0"
description = "IMU Visualization and Inverse Kinematics API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "quaternion>=2022.4.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.1",
    "Pillow>=10.0.0",
    # Database
    "supabase>=2.0.0",
    "sqlalchemy>=2.0.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    # Background tasks
    "celery>=5.3.0",
    "redis>=5.0.0",
    # Monitoring
    "sentry-sdk[fastapi]>=1.40.0",
    "structlog>=23.2.0",
    # Email
    "fastapi-mail>=1.4.0",
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"] 